Added support to run tests with custom Dockerfile
In order to support Kong EE we should be able to run custom containers
to our tests.

To achieve this, it's now allowed to provide a custom Dockerfile to be
used to build and run the Kong container.

Some notes:

- For now, it's only possible to set the desired Dockerfile using the
KONG_DOCKERFILE_PATH env var

- In order to avoid issues with Kong CE users, we added tags to the
tests, segregating them among "all", "community" and "enterprise", and,
therefore, allowing us to run each "kind" of test sepparately. This
means that the `make test` behavior remains the same and there are now 2
new commands: `make test-all` and `make test-enterprise`.

All of this was made necessary so we could write and run tests for
Kong's "Workspace" feature, which is only available on the Enterprise
version.